flatten plasma client definition
